+++ src/command_modules/azure-cli-batch/azure/cli/command_modules/batch/tests/latest/test_batch_mgmt_commands.py
@@ -3,27 +3,29 @@
 # Licensed under the MIT License. See License.txt in the project root for license information.
 # --------------------------------------------------------------------------------------------
 
-import tempfile
-import os
 import mock
+import os
+import tempfile
+import time
 
 from knack.util import CLIError
 from azure.cli.testsdk import TestCli
 from azure.cli.testsdk import (
     ScenarioTest, ResourceGroupPreparer, LiveScenarioTest)
 from .batch_preparers import BatchAccountPreparer, BatchScenarioMixin
-from azure.mgmt.keyvault.models import SecretPermissions, KeyPermissions
-
-# Key Vault permissions
-ALL_SECRET_PERMISSIONS = ' '.join([perm.value for perm in SecretPermissions])
-ALL_KEY_PERMISSIONS = ' '.join([perm.value for perm in KeyPermissions])
-
 
 # TODO: Convert back to ScenarioTest and re-record when issue #5142 is fixed.
 class BatchMgmtScenarioTests(LiveScenarioTest):  # pylint: disable=too-many-instance-attributes
 
     @ResourceGroupPreparer(location='northeurope')
     def test_batch_account_cmd(self, resource_group):
+        
+        SecretPermissions = get_sdk(self.cli_ctx, ResourceType.MGMT_KEYVAULT, 'models.key_vault_management_client_enums#SecretPermissions')
+        KeyPermissions = get_sdk(self.cli_ctx, ResourceType.MGMT_KEYVAULT, 'models.key_vault_management_client_enums#KeyPermissions')
+        ALL_SECRET_PERMISSIONS = ' '.join([perm.value for perm in SecretPermissions])
+        ALL_KEY_PERMISSIONS = ' '.join([perm.value for perm in KeyPermissions])
+
+        
         self.kwargs.update({
             'rg': resource_group,
             'str_n': 'clibatchteststorage1',
@@ -62,9 +64,7 @@ class BatchMgmtScenarioTests(LiveScenarioTest):  # pylint: disable=too-many-inst
             self.check('location', '{loc}'),
             self.check('resourceGroup', '{rg}')])
 
-        if self.is_live or self.in_recording:
-            import time
-            time.sleep(100)
+        time.sleep(100)
 
         # test create account with BYOS
         self.cmd('batch account create -g {rg} -n {byos_n} -l {byos_l} --keyvault {kv}').assert_with_checks([