+++ src/command_modules/azure-cli-sql/azure/cli/command_modules/sql/_validators.py
@@ -73,29 +73,6 @@ def create_args_for_complex_type(arg_ctx, dest, model_type, arguments):
 
 
 ###############################################
-#                sql db                       #
-###############################################
-
-
-def validate_elastic_pool_id(cmd, namespace):
-    from msrestazure.tools import resource_id, is_valid_resource_id
-    from azure.cli.core.commands.client_factory import get_subscription_id
-
-    # If elastic_pool_id is specified but it is not a valid resource id,
-    # then assume that user specified elastic pool name which we need to
-    # convert to elastic pool id.
-    if namespace.elastic_pool_id and not is_valid_resource_id(namespace.elastic_pool_id):
-        namespace.elastic_pool_id = resource_id(
-            subscription=get_subscription_id(cmd.cli_ctx),
-            resource_group=namespace.resource_group_name,
-            namespace='Microsoft.Sql',
-            type='servers',
-            name=namespace.server_name,
-            child_type_1='elasticPools',
-            child_name_1=namespace.elastic_pool_id)
-
-
-###############################################
 #                sql server vnet-rule         #
 ###############################################