+++ src/command_modules/azure-cli-vm/azure/cli/command_modules/vm/disk_encryption.py
@@ -584,10 +584,10 @@ def _verify_keyvault_good_for_encryption(cli_ctx, disk_vault_id, kek_vault_id, v
             raise CLIError(msg)
 
     from azure.cli.core.commands.client_factory import get_mgmt_service_client
-    from azure.mgmt.keyvault import KeyVaultManagementClient
+    from azure.cli.core.profiles import ResourceType
     from msrestazure.tools import parse_resource_id
 
-    client = get_mgmt_service_client(cli_ctx, KeyVaultManagementClient).vaults
+    client = get_mgmt_service_client(cli_ctx, ResourceType.MGMT_KEYVAULT).vaults
     disk_vault_resource_info = parse_resource_id(disk_vault_id)
     key_vault = client.get(disk_vault_resource_info['resource_group'], disk_vault_resource_info['name'])