+++ src/command_modules/azure-cli-keyvault/HISTORY.rst
@@ -2,6 +2,11 @@
 
 Release History
 ===============
+2.1.0
+++++++
+* support KV mgmt multi-api version
+* support KV data plane multi-api version
+
 2.0.24
 ++++++
 * adding commands for managing storage accounts and sas-definitions